requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
authlib==1.6.0
azure-core==1.34.0
azure-storage-blob==12.25.1
backoff==2.2.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.2.3
gremlinpython==3.7.3
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-pydantic==0.6
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-groq==0.3.5
langchain-text-splitters==0.3.8
langsmith==0.4.4
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.10.1
mcp-use==1.3.3
mdurl==0.1.2
multidict==6.6.3
mypy-extensions==1.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.3.0
posthog==6.0.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
scarf-sdk==0.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
